refactor(phase-2j): extract OpenProjectUserService
Continues the openproject_client.py god-class decomposition (ADR-002).
All seven user-related helpers — full-list reads, single-user lookup,
email lookup, batched id/email reads, and the avatar settings + upload
pair — move into a focused ``OpenProjectUserService`` exposed as
``self.users``. ``OpenProjectClient`` keeps thin delegators for each
method so existing callers work unchanged.

Methods moved
-------------
- ``get_users``
- ``get_user``
- ``get_user_by_email``
- ``batch_get_users_by_ids``
- ``batch_get_users_by_emails`` (the ``@batch_idempotent`` decorator
  travels with the moved method so cache keys are identical)
- ``ensure_local_avatars_enabled``
- ``set_user_avatar`` (lazy-imports ``escape_ruby_single_quoted`` to
  avoid the service ↔ client cycle)

Caches stay on the client
-------------------------
``_users_cache``, ``_users_cache_time``, and ``_users_by_email_cache``
remain on ``OpenProjectClient.__init__`` because other client paths
(``find_record``, project/role helpers) read them too. Service methods
reach through ``self._client.<cache>``.

Numbers
-------
- ``openproject_client.py``: 5,375 → 5,103 LOC (-272)
- ``openproject_user_service.py``: 0 → 414 LOC (new)
- Cumulative across phases 2a–2j: ``openproject_client.py``
  7,342 → 5,103 LOC (-2,239, -30.5%)

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean (102 files)
- ruff check / ruff format: clean
